+++ package.json
@@ -32,6 +32,7 @@
     "react-swipeable-views": "^0.14.0",
     "styled-components": "^5.3.5",
     "swiper": "^8.3.2",
+    "tailwind-scrollbar-hide": "^1.1.7",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
