+++ src/index.css
@@ -7,7 +7,7 @@
     padding: 0;
     box-sizing: border-box;
 
-
+    
 }
 
 html {
@@ -20,4 +20,8 @@ html {
     -khtml-user-select: none;
     -moz-user-select: none;
     -ms-user-select: none;
+}
+
+::-webkit-scrollbar {
+    display: none;
 }
